Beta - v0.6.0
# What's New

## Add Module - Multiplayer Sync
Synchronizes and disables all players until the required amount of players are fully loaded and spawned, or a server timeout occurs. Useful for waiting for everyone before playing cinematics or scripts that you want everyone to experience, regardless of load times.

# Patch

## Move embedded effects functions from module functions to addon: _Effects_
This is mainly for code organization. Should not affect anyone.