Add Feature - Communication Jammer (#58)
# Add Feature - Communication Jammer

## What's New

### Communication Jammer
This module will turn an object (or area if not synced to an object) into a jamming zone. As the player enters a jamming zone, their map and GPS will begin to decrease in visibility, changing with the distance from the jamming source. Not only will their maps decrease in visibility, but all chat messages sent to the player will become increasingly garbled in text becoming virtually unreadable at the jammer source.

Note, if the jamming source is an object, the jammer WILL MOVE with the object. This means you can create mobile jamming sources. If multiple jamming sources are present, the higher jamming amount will be applied.

Future:
Currently working on a ACRE2 compatibility patch that will also scramble radio communications.